bug & logic fix
