+++ tg_bot/tasks/dialog_states/collection.py
@@ -1,6 +1,9 @@
 from aiogram.dispatcher.filters.state import State, StatesGroup
 
 
+DIALOG_NOT_AVAILABLE_STATE = State()
+
+
 class FlagsToTextCreating(StatesGroup):  # TODO: add state when bot already creating answer
     web_search = State()
