+++ tg_bot/tasks/dialog_states/processors.py
@@ -1,7 +1,16 @@
 from ...base import DISPATCHER
 from .collection import DialogStates
+from .dialog_allowance import DialogNotAvailable
 from aiogram.types import Message
 from aiogram.dispatcher import FSMContext
+
+
+@DISPATCHER.message_handler(DialogNotAvailable())
+async def send_answer_if_dialog_not_available(message: Message):
+    await message.answer(
+        text="Извините, вы *не можете* пользоваться этим ботом, обратитесь к человеку, который дал вам ссылку 🤐",
+        parse_mode="Markdown",
+    )
 
 
 @DISPATCHER.message_handler(state=DialogStates.answer_is_creating)
