+++ tg_bot/tasks/misc.py
@@ -1,4 +1,5 @@
 from ..base import DISPATCHER
+from ai import get_user
 from aiogram.types import CallbackQuery, Message
 
 
@@ -35,11 +36,12 @@ HELP_ANSWER_TEXT = """Добро пожаловать в бота, {full_name}!
 
 @DISPATCHER.message_handler(commands=["start", "help"], state="*")
 async def handle_start_help_command(message: Message):
-    # TODO: replace message.from_user.full_name to database.user.full_name
-    await message.reply(HELP_ANSWER_TEXT.format(full_name=message.from_user.full_name), parse_mode="Markdown")
+    await message.reply(
+        text=HELP_ANSWER_TEXT.format(full_name=get_user(message.from_user.id).user_full_name), parse_mode="Markdown"
+    )
 
 
-@DISPATCHER.callback_query_handler(state="TODO")
+@DISPATCHER.callback_query_handler(text="TODO", state="*")
 async def handle_todo_callback(callback: CallbackQuery):
     await callback.answer("👷Данный блок находится в разработке...", show_alert=True)
